virt-launcher: move migration-related tests to proper file
Move tests out of manager_test.go and into
live-migration-source_test.go, where the implementations live.
