+++ pkg/virt-launcher/virtwrap/manager_test.go
@@ -23,7 +23,6 @@ import (
 	"crypto/sha256"
 	_ "embed"
 	"encoding/base64"
-	"encoding/json"
 	"encoding/xml"
 	"fmt"
 	"os"
@@ -36,12 +35,10 @@ import (
 	. "github.com/onsi/gomega"
 	"go.uber.org/mock/gomock"
 	"libvirt.org/go/libvirt"
-	"libvirt.org/go/libvirtxml"
 
 	"kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/disksource"
 	"kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/network"
 
-	k8sv1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
@@ -2235,49 +2232,6 @@ var _ = Describe("Manager", func() {
 			migration, _ := metadataCache.Migration.Load()
 			Expect(migration).To(Equal(startupMigrationMetadata))
 		})
-		It("should correctly collect a list of disks for migration", func() {
-			_true := true
-			vmi := newVMI(testNamespace, testVmName)
-			vmi.Spec.Volumes = []v1.Volume{
-				{
-					Name: "myvolume",
-					VolumeSource: v1.VolumeSource{
-						PersistentVolumeClaim: &v1.PersistentVolumeClaimVolumeSource{PersistentVolumeClaimVolumeSource: k8sv1.PersistentVolumeClaimVolumeSource{
-							ClaimName: "testblock",
-						}},
-					},
-				},
-				{
-					Name: "myvolume1",
-					VolumeSource: v1.VolumeSource{
-						Ephemeral: &v1.EphemeralVolumeSource{
-							PersistentVolumeClaim: &k8sv1.PersistentVolumeClaimVolumeSource{
-								ClaimName: "testclaim",
-							},
-						},
-					},
-				},
-				{
-					Name: "myvolumehost",
-					VolumeSource: v1.VolumeSource{
-						HostDisk: &v1.HostDisk{
-							Path:     "/var/run/kubevirt-private/vmi-disks/volume3/disk.img",
-							Type:     v1.HostDiskExistsOrCreate,
-							Capacity: resource.MustParse("1Gi"),
-							Shared:   &_true,
-						},
-					},
-				},
-			}
-			userData := "fake\nuser\ndata\n"
-			networkData := "FakeNetwork"
-			addCloudInitDisk(vmi, userData, networkData)
-
-			mockLibvirt.DomainEXPECT().GetXMLDesc(libvirt.DomainXMLFlags(0)).Return(embedMigrationDomain, nil)
-
-			copyDisks := getDiskTargetsForMigration(mockLibvirt.VirtDomain, vmi)
-			Expect(copyDisks).Should(ConsistOf("vdb", "vdd"))
-		})
 		AfterEach(func() {
 			ip.GetLoopbackAddress = funcPreviousValue
 		})
@@ -2307,52 +2261,6 @@ var _ = Describe("Manager", func() {
 			Entry("paused", libvirt.DOMAIN_PAUSED),
 		)
 	})
-	DescribeTable("check migration flags",
-		func(migrationType string) {
-			isBlockMigration := migrationType == "block"
-			isVmiPaused := migrationType == "paused"
-
-			options := &cmdclient.MigrationOptions{
-				UnsafeMigration:   migrationType == "unsafe",
-				AllowAutoConverge: migrationType == "autoConverge",
-				AllowPostCopy:     migrationType == "postCopy",
-			}
-
-			shouldConfigureParallel, parallelMigrationThreads := shouldConfigureParallelMigration(options)
-			if shouldConfigureParallel {
-				options.ParallelMigrationThreads = virtpointer.P(uint(parallelMigrationThreads))
-			}
-
-			flags := generateMigrationFlags(isBlockMigration, isVmiPaused, options)
-			expectedMigrateFlags := libvirt.MIGRATE_LIVE | libvirt.MIGRATE_PEER2PEER | libvirt.MIGRATE_PERSIST_DEST
-
-			if isBlockMigration {
-				expectedMigrateFlags |= libvirt.MIGRATE_NON_SHARED_INC
-			} else if migrationType == "unsafe" {
-				expectedMigrateFlags |= libvirt.MIGRATE_UNSAFE
-			}
-			if options.AllowAutoConverge {
-				expectedMigrateFlags |= libvirt.MIGRATE_AUTO_CONVERGE
-			}
-			if migrationType == "postCopy" {
-				expectedMigrateFlags |= libvirt.MIGRATE_POSTCOPY
-			}
-			if migrationType == "paused" {
-				expectedMigrateFlags |= libvirt.MIGRATE_PAUSED
-			}
-			if shouldConfigureParallel {
-				expectedMigrateFlags |= libvirt.MIGRATE_PARALLEL
-			}
-			Expect(flags).To(Equal(expectedMigrateFlags), "libvirt migration flags are not set as expected")
-		},
-		Entry("with block migration", "block"),
-		Entry("without block migration", "live"),
-		Entry("unsafe migration", "unsafe"),
-		Entry("migration auto converge", "autoConverge"),
-		Entry("migration using postcopy", "postCopy"),
-		Entry("migration of paused vmi", "paused"),
-	)
-
 	DescribeTable("on successful list all domains",
 		func(state libvirt.DomainState, kubevirtState api.LifeCycle, libvirtReason int, kubevirtReason api.StateChangeReason) {
 
@@ -3536,228 +3444,6 @@ var _ = Describe("getUpdatedDisks", func() {
 	)
 })
 
-var _ = Describe("migratableDomXML", func() {
-	var ctrl *gomock.Controller
-	var mockLibvirt *testing.Libvirt
-	BeforeEach(func() {
-		ctrl = gomock.NewController(GinkgoT())
-		mockLibvirt = testing.NewLibvirt(ctrl)
-	})
-	It("should parse the XML with the metadata", func() {
-		domXML := `<domain type="kvm" id="1">
-  <name>kubevirt</name>
-  <metadata>
-    <kubevirt xmlns="http://kubevirt.io">
-    </kubevirt>
-   </metadata>
-</domain>`
-		expectedXML := `<domain type="kvm" id="1">
-  <name>kubevirt</name>
-  <metadata>
-    <kubevirt xmlns="http://kubevirt.io">
-    </kubevirt>
-   </metadata>
-</domain>`
-		vmi := newVMI("testns", "kubevirt")
-		mockLibvirt.DomainEXPECT().GetXMLDesc(libvirt.DOMAIN_XML_MIGRATABLE).MaxTimes(1).Return(domXML, nil)
-		domSpec := &api.DomainSpec{}
-		Expect(xml.Unmarshal([]byte(domXML), domSpec)).To(Succeed())
-		newXML, err := migratableDomXML(mockLibvirt.VirtDomain, vmi, domSpec)
-		Expect(err).ToNot(HaveOccurred())
-		Expect(newXML).To(Equal(expectedXML))
-	})
-	It("should change CPU pinning according to migration metadata", func() {
-		domXML := `<domain type="kvm" id="1">
-  <name>kubevirt</name>
-  <vcpu placement="static">2</vcpu>
-  <cputune>
-    <vcpupin vcpu="0" cpuset="4"></vcpupin>
-    <vcpupin vcpu="1" cpuset="5"></vcpupin>
-  </cputune>
-</domain>`
-		// migratableDomXML() removes the migration block but not its ident, which is its own token, hence the blank line below
-		expectedXML := `<domain type="kvm" id="1">
-  <name>kubevirt</name>
-  <vcpu placement="static">2</vcpu>
-  <cputune>
-    <vcpupin vcpu="0" cpuset="6"></vcpupin>
-    <vcpupin vcpu="1" cpuset="7"></vcpupin>
-  </cputune>
-  <cpu>
-    <topology sockets="1" cores="2" threads="1"></topology>
-  </cpu>
-</domain>`
-
-		By("creating a VMI with dedicated CPU cores")
-		vmi := newVMI("testns", "kubevirt")
-		vmi.Spec.Domain.CPU = &v1.CPU{
-			Cores:                 2,
-			DedicatedCPUPlacement: true,
-		}
-
-		By("making up a target topology")
-		topology := &cmdv1.Topology{NumaCells: []*cmdv1.Cell{{
-			Id: 0,
-			Cpus: []*cmdv1.CPU{
-				{
-					Id:       6,
-					Siblings: []uint32{6},
-				},
-				{
-					Id:       7,
-					Siblings: []uint32{7},
-				},
-			},
-		}}}
-		targetNodeTopology, err := json.Marshal(topology)
-		Expect(err).NotTo(HaveOccurred(), "failed to marshall the topology")
-
-		By("saving that topology in the migration state of the VMI")
-		vmi.Status.MigrationState = &v1.VirtualMachineInstanceMigrationState{
-			TargetCPUSet:       []int{6, 7},
-			TargetNodeTopology: string(targetNodeTopology),
-		}
-
-		By("generated the domain XML for a migration to that target")
-		mockLibvirt.DomainEXPECT().GetXMLDesc(libvirt.DOMAIN_XML_MIGRATABLE).MaxTimes(1).Return(domXML, nil)
-		domSpec := &api.DomainSpec{}
-		Expect(xml.Unmarshal([]byte(domXML), domSpec)).To(Succeed())
-		Expect(domSpec.VCPU).NotTo(BeNil())
-		Expect(domSpec.CPUTune).NotTo(BeNil())
-		newXML, err := migratableDomXML(mockLibvirt.VirtDomain, vmi, domSpec)
-		Expect(err).ToNot(HaveOccurred(), "failed to generate target domain XML")
-
-		By("ensuring the generated XML is accurate")
-		Expect(newXML).To(Equal(expectedXML), "the target XML is not as expected")
-	})
-	DescribeTable("slices section", func(domXML string) {
-		retDiskSize := func(disk *libvirtxml.DomainDisk) (int64, error) {
-			return 2028994560, nil
-		}
-		getDiskVirtualSizeFunc = retDiskSize
-		const (
-			volName       = "datavolumedisk1"
-			sourcePvcName = "src-pvc"
-			destPvcName   = "dst-pvc"
-		)
-		expectedXML := `<domain type="kvm" id="1">
-  <name>kubevirt</name>
-  <devices>
-    <disk type="file" device="disk" model="virtio-non-transitional">
-      <driver name="qemu" type="raw" cache="none" error_policy="stop" discard="unmap"></driver>
-      <source file="/var/run/kubevirt-private/vmi-disks/datavolumedisk1/disk.img" index="1">
-        <slices>
-          <slice type="storage" offset="0" size="2028994560"></slice>
-        </slices>
-      </source>
-      <backingStore></backingStore>
-      <target dev="vda" bus="virtio"></target>
-      <alias name="ua-datavolumedisk1"></alias>
-      <address type="pci" domain="0x0000" bus="0x07" slot="0x00" function="0x0"></address>
-    </disk>
-  </devices>
-</domain>`
-		vmi := newVMI("testns", "kubevirt")
-		vmi.Spec.Volumes = append(vmi.Spec.Volumes,
-			v1.Volume{
-				Name: volName,
-				VolumeSource: v1.VolumeSource{
-					DataVolume: &v1.DataVolumeSource{
-						Name: sourcePvcName,
-					},
-				},
-			})
-		vmi.Status.MigratedVolumes = []v1.StorageMigratedVolumeInfo{
-			{
-				VolumeName: volName,
-				SourcePVCInfo: &v1.PersistentVolumeClaimInfo{
-					ClaimName:  sourcePvcName,
-					VolumeMode: virtpointer.P(k8sv1.PersistentVolumeFilesystem),
-				},
-				DestinationPVCInfo: &v1.PersistentVolumeClaimInfo{
-					ClaimName:  destPvcName,
-					VolumeMode: virtpointer.P(k8sv1.PersistentVolumeFilesystem),
-				},
-			},
-		}
-		mockLibvirt.DomainEXPECT().GetXMLDesc(libvirt.DOMAIN_XML_MIGRATABLE).MaxTimes(1).Return(domXML, nil)
-		domSpec := &api.DomainSpec{}
-		Expect(xml.Unmarshal([]byte(domXML), domSpec)).To(Succeed())
-		newXML, err := migratableDomXML(mockLibvirt.VirtDomain, vmi, domSpec)
-		Expect(err).ToNot(HaveOccurred())
-		Expect(newXML).To(Equal(expectedXML))
-	},
-		Entry("add slices section", `<domain type="kvm" id="1">
-  <name>kubevirt</name>
-  <devices>
-    <disk type='file' device='disk' model='virtio-non-transitional'>
-      <driver name='qemu' type='raw' cache='none' error_policy='stop' discard='unmap'/>
-      <source file='/var/run/kubevirt-private/vmi-disks/datavolumedisk1/disk.img' index='1'/>
-      <backingStore/>
-      <target dev='vda' bus='virtio'/>
-      <alias name='ua-datavolumedisk1'/>
-      <address type='pci' domain='0x0000' bus='0x07' slot='0x00' function='0x0'/>
-    </disk>
-  </devices>
-</domain>`),
-		Entry("slices section already set", `<domain type="kvm" id="1">
-  <name>kubevirt</name>
-  <devices>
-    <disk type='file' device='disk' model='virtio-non-transitional'>
-      <driver name='qemu' type='raw' cache='none' error_policy='stop' discard='unmap'/>
-      <source file='/var/run/kubevirt-private/vmi-disks/datavolumedisk1/disk.img' index='1'>
-        <slices>
-          <slice type='storage' offset='0' size='2028994560'></slice>
-        </slices>
-      </source>
-      <backingStore/>
-      <target dev='vda' bus='virtio'/>
-      <alias name='ua-datavolumedisk1'/>
-      <address type='pci' domain='0x0000' bus='0x07' slot='0x00' function='0x0'/>
-    </disk>
-  </devices>
-</domain>`),
-	)
-	It("should generate correct xml for user data for copied disks during the migration", func() {
-		domXML := `<domain type="kvm" id="1">
-  <name>kubevirt</name>
-  <devices>
-    <disk type='file' device='disk' model='virtio-non-transitional'>
-      <driver name='qemu' type='raw' cache='none' error_policy='stop' discard='unmap'/>
-      <source file='/var/run/kubevirt-ephemeral-disks/cloud-init-data/default/vm-dv/noCloud.iso' index='1'/>
-      <backingStore/>
-      <target dev='vda' bus='virtio'/>
-      <alias name='ua-cloudinitdisk'/>
-      <address type='pci' domain='0x0000' bus='0x07' slot='0x00' function='0x0'/>
-    </disk>
-  </devices>
-</domain>`
-		expectedXML := `<domain type="kvm" id="1">
-  <name>kubevirt</name>
-  <devices>
-    <disk type="file" device="disk" model="virtio-non-transitional">
-      <driver name="qemu" type="raw" cache="none" error_policy="stop" discard="unmap"></driver>
-      <source file="/var/run/kubevirt-ephemeral-disks/cloud-init-data/default/vm-dv/noCloud.iso" index="1"></source>
-      <backingStore></backingStore>
-      <target dev="vda" bus="virtio"></target>
-      <alias name="ua-cloudinitdisk"></alias>
-      <address type="pci" domain="0x0000" bus="0x07" slot="0x00" function="0x0"></address>
-    </disk>
-  </devices>
-</domain>`
-		vmi := newVMI("testns", "kubevirt")
-		userData := "fake\nuser\ndata\n"
-		networkData := "FakeNetwork"
-		addCloudInitDisk(vmi, userData, networkData)
-		mockLibvirt.DomainEXPECT().GetXMLDesc(libvirt.DOMAIN_XML_MIGRATABLE).MaxTimes(1).Return(domXML, nil)
-		domSpec := &api.DomainSpec{}
-		Expect(xml.Unmarshal([]byte(domXML), domSpec)).To(Succeed())
-		newXML, err := migratableDomXML(mockLibvirt.VirtDomain, vmi, domSpec)
-		Expect(err).ToNot(HaveOccurred())
-		Expect(newXML).To(Equal(expectedXML))
-	})
-})
-
 var _ = Describe("Manager helper functions", func() {
 
 	Context("getVMIEphemeralDisksTotalSize", func() {
@@ -3927,249 +3613,6 @@ var _ = Describe("Manager helper functions", func() {
 					},
 				}
 			}),
-		)
-	})
-
-	Context("configureLocalDiskToMigrate", func() {
-		const (
-			testvol = "test"
-			src     = "src"
-			dst     = "dst"
-		)
-
-		fsMode := k8sv1.PersistentVolumeFilesystem
-		blockMode := k8sv1.PersistentVolumeBlock
-		infoFs := v1.PersistentVolumeClaimInfo{
-			ClaimName:  src,
-			VolumeMode: &fsMode,
-		}
-		infoBlock := v1.PersistentVolumeClaimInfo{
-			ClaimName:  src,
-			VolumeMode: &blockMode,
-		}
-
-		createDomWithFsImage := func(name string) *libvirtxml.Domain {
-			return &libvirtxml.Domain{
-				Devices: &libvirtxml.DomainDeviceList{
-					Disks: []libvirtxml.DomainDisk{
-						{
-							Source: &libvirtxml.DomainDiskSource{
-								File: &libvirtxml.DomainDiskSourceFile{
-									File: getFsImagePath(name),
-								},
-							},
-							Alias: &libvirtxml.DomainAlias{
-								Name: fmt.Sprintf("ua-%s", name),
-							},
-						},
-					},
-				},
-			}
-		}
-		createDomWithBlock := func(name string) *libvirtxml.Domain {
-			return &libvirtxml.Domain{
-				Devices: &libvirtxml.DomainDeviceList{
-					Disks: []libvirtxml.DomainDisk{
-						{
-							Source: &libvirtxml.DomainDiskSource{
-								Block: &libvirtxml.DomainDiskSourceBlock{
-									Dev: getBlockPath(name),
-								},
-							},
-							Alias: &libvirtxml.DomainAlias{
-								Name: fmt.Sprintf("ua-%s", name),
-							},
-						},
-					},
-				},
-			}
-		}
-		volPVC := v1.Volume{
-			Name: testvol,
-			VolumeSource: v1.VolumeSource{
-				PersistentVolumeClaim: &v1.PersistentVolumeClaimVolumeSource{
-					PersistentVolumeClaimVolumeSource: k8sv1.PersistentVolumeClaimVolumeSource{
-						ClaimName: src,
-					},
-				},
-			},
-		}
-		volDV := v1.Volume{
-			Name: testvol,
-			VolumeSource: v1.VolumeSource{
-				DataVolume: &v1.DataVolumeSource{
-					Name: src,
-				},
-			},
-		}
-		volHostDisk := v1.Volume{
-			Name: testvol,
-			VolumeSource: v1.VolumeSource{
-				HostDisk: &v1.HostDisk{
-					Path: getFsImagePath(testvol),
-				},
-			},
-		}
-
-		DescribeTable("replace filesystem and block migrated volumes", func(isSrcBlock, isDstBlock bool, vol v1.Volume) {
-			retDiskSize := func(disk *libvirtxml.DomainDisk) (int64, error) {
-				return 2028994560, nil
-			}
-			getDiskVirtualSizeFunc = retDiskSize
-			var dom *libvirtxml.Domain
-			vmi := &v1.VirtualMachineInstance{
-				Spec: v1.VirtualMachineInstanceSpec{
-					Volumes: []v1.Volume{vol},
-				},
-				Status: v1.VirtualMachineInstanceStatus{
-					MigratedVolumes: []v1.StorageMigratedVolumeInfo{
-						{
-							VolumeName: testvol,
-						},
-					},
-					VolumeStatus: []v1.VolumeStatus{
-						{
-							Name: testvol,
-							PersistentVolumeClaimInfo: &v1.PersistentVolumeClaimInfo{
-								ClaimName: src,
-							},
-						},
-					},
-				},
-			}
-			if isSrcBlock {
-				vmi.Status.MigratedVolumes[0].SourcePVCInfo = &infoBlock
-				dom = createDomWithBlock(testvol)
-			} else {
-				vmi.Status.MigratedVolumes[0].SourcePVCInfo = &infoFs
-				dom = createDomWithFsImage(testvol)
-			}
-			if isDstBlock {
-				vmi.Status.MigratedVolumes[0].DestinationPVCInfo = &infoBlock
-			} else {
-				vmi.Status.MigratedVolumes[0].DestinationPVCInfo = &infoFs
-			}
-
-			err := configureLocalDiskToMigrate(dom, vmi)
-			Expect(err).ToNot(HaveOccurred())
-
-			if isDstBlock {
-				Expect(dom.Devices.Disks[0].Source.File).To(BeNil())
-				Expect(dom.Devices.Disks[0].Source.Block).NotTo(BeNil())
-				Expect(dom.Devices.Disks[0].Source.Block.Dev).To(Equal(getBlockPath(testvol)))
-
-			} else {
-				Expect(dom.Devices.Disks[0].Source.Block).To(BeNil())
-				Expect(dom.Devices.Disks[0].Source.File).NotTo(BeNil())
-				Expect(dom.Devices.Disks[0].Source.File.File).To(Equal(getFsImagePath(testvol)))
-			}
-		},
-			Entry("filesystem source and destination", false, false, volPVC),
-			Entry("filesystem source and block destination", false, true, volPVC),
-			Entry("block source and filesystem destination", true, false, volPVC),
-			Entry("block source and destination", true, true, volPVC),
-			Entry("filesystem source and block destination with DV", false, true, volDV),
-			Entry("block source and filesystem destination with DV", true, false, volDV),
-			Entry("filesystem source and block destination with hostdisks", false, true, volHostDisk),
-			Entry("block source and filesystem destination with hostdisks", true, false, volHostDisk),
-		)
-
-		DescribeTable("replace filesystem and block migrated volumes with CBT overlay", func(isSrcBlock, isDstBlock bool) {
-			retDiskSize := func(disk *libvirtxml.DomainDisk) (int64, error) {
-				return 2028994560, nil
-			}
-			getDiskVirtualSizeFunc = retDiskSize
-
-			cbtOverlayPath := "/var/lib/libvirt/qemu/cbt/" + testvol + ".qcow2"
-			var backendSrc *libvirtxml.DomainDiskSource
-			if isSrcBlock {
-				backendSrc = &libvirtxml.DomainDiskSource{Block: &libvirtxml.DomainDiskSourceBlock{Dev: getBlockPath(testvol)}}
-			} else {
-				backendSrc = &libvirtxml.DomainDiskSource{File: &libvirtxml.DomainDiskSourceFile{File: getFsImagePath(testvol)}}
-			}
-			dom := &libvirtxml.Domain{
-				Devices: &libvirtxml.DomainDeviceList{
-					Disks: []libvirtxml.DomainDisk{
-						{
-							Source: &libvirtxml.DomainDiskSource{
-								File:      &libvirtxml.DomainDiskSourceFile{File: cbtOverlayPath},
-								DataStore: &libvirtxml.DomainDiskDataStore{Source: backendSrc},
-							},
-							Alias: &libvirtxml.DomainAlias{Name: fmt.Sprintf("ua-%s", testvol)},
-						},
-					},
-				},
-			}
-			vmi := &v1.VirtualMachineInstance{
-				Spec: v1.VirtualMachineInstanceSpec{
-					Volumes: []v1.Volume{volPVC},
-				},
-				Status: v1.VirtualMachineInstanceStatus{
-					MigratedVolumes: []v1.StorageMigratedVolumeInfo{
-						{
-							VolumeName: testvol,
-						},
-					},
-					VolumeStatus: []v1.VolumeStatus{
-						{
-							Name:                      testvol,
-							PersistentVolumeClaimInfo: &v1.PersistentVolumeClaimInfo{ClaimName: src},
-						},
-					},
-				},
-			}
-			if isSrcBlock {
-				vmi.Status.MigratedVolumes[0].SourcePVCInfo = &infoBlock
-			} else {
-				vmi.Status.MigratedVolumes[0].SourcePVCInfo = &infoFs
-			}
-			if isDstBlock {
-				vmi.Status.MigratedVolumes[0].DestinationPVCInfo = &infoBlock
-			} else {
-				vmi.Status.MigratedVolumes[0].DestinationPVCInfo = &infoFs
-			}
-
-			err := configureLocalDiskToMigrate(dom, vmi)
-			Expect(err).ToNot(HaveOccurred())
-
-			diskSrc := dom.Devices.Disks[0].Source
-			Expect(diskSrc.File).NotTo(BeNil())
-			Expect(diskSrc.File.File).To(Equal(cbtOverlayPath))
-			Expect(diskSrc.DataStore).NotTo(BeNil())
-			Expect(diskSrc.DataStore.Source).NotTo(BeNil())
-			if isDstBlock {
-				Expect(diskSrc.DataStore.Source.File).To(BeNil())
-				Expect(diskSrc.DataStore.Source.Block).NotTo(BeNil())
-				Expect(diskSrc.DataStore.Source.Block.Dev).To(Equal(getBlockPath(testvol)))
-			} else {
-				Expect(diskSrc.DataStore.Source.Block).To(BeNil())
-				Expect(diskSrc.DataStore.Source.File).NotTo(BeNil())
-				Expect(diskSrc.DataStore.Source.File.File).To(Equal(getFsImagePath(testvol)))
-			}
-		},
-			Entry("filesystem source and block destination", false, true),
-			Entry("block source and filesystem destination", true, false),
-			Entry("filesystem source and filesystem destination", false, false),
-			Entry("block source and block destination", true, true),
-		)
-	})
-
-	Context("shouldConfigureParallelMigration", func() {
-		DescribeTable("should not configure parallel migration", func(options *cmdclient.MigrationOptions) {
-			shouldConfigure, _ := shouldConfigureParallelMigration(options)
-			Expect(shouldConfigure).To(BeFalse())
-		},
-			Entry("with nil options", nil),
-			Entry("with nil migration threads", &cmdclient.MigrationOptions{ParallelMigrationThreads: nil}),
-			Entry("with nil migration threads and post-copy allowed", &cmdclient.MigrationOptions{ParallelMigrationThreads: nil, AllowPostCopy: true}),
-		)
-
-		DescribeTable("should configure parallel migration", func(options *cmdclient.MigrationOptions) {
-			shouldConfigure, _ := shouldConfigureParallelMigration(options)
-			Expect(shouldConfigure).To(BeTrue())
-		},
-			Entry("with non-nil migration threads and post-copy not allowed", &cmdclient.MigrationOptions{ParallelMigrationThreads: virtpointer.P(uint(3)), AllowPostCopy: false}),
-			Entry("with non-nil migration threads and post-copy allowed", &cmdclient.MigrationOptions{ParallelMigrationThreads: virtpointer.P(uint(3)), AllowPostCopy: true}),
 		)
 	})
 })
